Extract governance-lib and split commands.rs into different files
