+++ governance-lib/src/lib.rs
@@ -1,0 +1,17 @@
+pub mod client;
+
+pub mod realm;
+pub mod governance;
+pub mod proposal;
+pub mod token_owner;
+
+pub mod addin_fixed_weights;
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        let result = 2 + 2;
+        assert_eq!(result, 4);
+    }
+}
